pyproject.toml
[tool.poetry]
name = "savingapi"
version = "0.1.0"
description = ""
authors = ["scifisatan <brawlershere@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pytz = "^2024.1"
supabase-py = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
